+++ src/test/java/org/mmarini/railways2/model/blocks/RightPlatformSwitchTest.java
@@ -31,7 +31,7 @@ package org.mmarini.railways2.model.blocks;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mmarini.Tuple2;
-import org.mmarini.railways2.model.geometry.EdgeBuilder;
+import org.mmarini.railways2.model.geometry.EdgeBuilderParams;
 import org.mmarini.railways2.model.geometry.Node;
 import org.mmarini.railways2.model.geometry.NodeBuilderParams;
 import org.mmarini.railways2.model.routes.Route;
@@ -92,7 +92,7 @@ class RightPlatformSwitchTest {
     void getEdgeBuilders() {
         // Given ...
         // When ...
-        List<EdgeBuilder> builders = block.getEdgeBuilders();
+        List<EdgeBuilderParams> builders = block.getEdgeParams();
 
         // Then ...
         assertThat(builders, hasSize(6));