+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Issue #1: Add building blocks
 - Issue #4: Immutable geometry, routes, trains
 - Issue #21: Add double switch
 - Issue #28: Simulator with rx4j### Changed